third_party/blink/web_tests/platform/generic/fast/forms/select/listbox-click-on-scrollbar-expected.txt

Click on a listbox scrollbar should not dispatch change evnet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select1.selectedIndex is -1
PASS changeEventCounter is 0


Dispatch a change evnet to prepare the next test.
A change event was dispatched.
PASS changeEventCounter is 1
Change the option list size, and click the scrollbar again.
PASS changeEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE

1
2
3
4
5
6
7
8
9
